requirements.txt
flask>=2.0.0
gunicorn>=20.0.0
python-dotenv
filelock
requests
